App SmartMammo/requirements.txt
streamlit
Pillow
transformers
torch
pydicom
numpy
albumentations
fpdf
torchvision
streamlit_drawable_canvas